Post-process published examples to get rid of MATLAB LaTeX errors.
Ugh... how annoying.  But there's not much else we can do unless we hack MATLAB
publish() directly using Mr. Hrothgar's changes, and we can't do that on our
workstations (since we don't have write permissions for the MATLAB install
directory).

The strip-mcode-errors.pl Perl script requires XML::LibXML, which comes
installed by default on our workstations.  For personal machines, it can be
acquired quickly and easilty through CPAN.  Alternatively, if someone wants to
convert this to Python (which I think ships an XML parser with its standard
library, so there are no external dependencies), that's cool too.  I only used
Perl because it's what I know, and therefore it's what I could use to get the
task done as quickly as possible.

If the script cannot run because someone does not have the requisite
dependency, make_example() should still work, but the error output will not be
stripped.