Track download link clicks using Google Analytics.
This uses jQuery to set up a handler that will send an event to Google
Analytics every time someone clicks the download link on the download
page.  Note that this will ONLY capture downloads made in this manner.
In particular, it does NOT pick up downloads

  - using the 2-line MATLAB script,

  - made directly from GitHub, or

  - from the MATLAB File Exchange.

I had to convert the hyperlink in download/index.md to raw HTML so I
could assign an ID to it for use with the JavaScript.  The JavaScript is
added to the page by adding a new block for page-specific JS in
_templates/_base.html and then creating a new template for the download
page that fills in this block.  The rest of the download page template
is inherited from _templates/basic.html.