Change some figure parameters in make_example.m.
In particular, we set DefaultAxesFontSize to 10 and turn off the bold
font in titles.  These seem to give more reasonable-looking results in
the world of R2014b and later.

Trouble with figure sizes in ode-nonlin/ChebopQuiver.m on Dr.
Birkisson's Mac suggests that there is more to do here and that the
changes may be screen-resolution-dependent.  But there's no time to deal
with these right at the moment.