Update instructions for subscribing to mailing list.
As far as I can tell, it's no longer possible to join a Google group
through the web interface without a Google account.  This is Not
Friendly Behavior on Google's part, but there's little we can do about
it aside from hosting the mailing list somewhere else, which we don't
want to do at this time.  Instead of just posting a link to the page on
Google groups, we now also give the instructions for joining if you
don't have a Google account, which requires you to send a few e-mails.

Also, added a line to the homepage advertising the mailing list with a
link to the subscription instructions.