Specify font size in pixels for plots in examples and the Guide in R2015b+.
I think the reason we've (or, at least, I've) been having so many issues
with font sizes when publishing examples recently is that in R2015b,
MATLAB became "DPI-aware".  As a result, MATLAB now supposedly renders a
font whose size is specified as 10 points at a true 10 points.  That is,
MATLAB knows the actual display resolution (in dots-per-inch) and
selects an appropriate number of pixels to draw the text at 10 points.
Previously, MATLAB would just assume the display resolution to be a
fixed, platform-dependent quantity (e.g., 96 dpi on Windows) and render
using that value, irrespective of the true display resolution.

Now, If a character in a 10 point font takes up, a certain number of
pixels at 96 dpi, it will take 50% more pixels to display at the same
size at 144 dpi, which, incidentally, is the resolution of my laptop's
display.  Because MATLAB code now does this adjustment automatically in
R2015b, making the upgrade makes all of the characters in our plots 50%
bigger on my machine when they get displayed.

Our problem is the combination of this new behavior with the fact that
our web graphics are of a fixed size _in pixels_, not real-world units
like points and inches.  Therefore, the plots stayed the same size
during the upgrade, while all of the characters got bigger.  The easiest
way to fix this is to specify the font size in pixels instead of points,
and that's what is done in this commit.  The value of 13.3333 that I've
chosen should correspond to a 10-point font at 96 dpi, since 96*(10/72)
= 13.3333...

I've opted to make these changes conditional on the currently running
version of MATLAB so that the old behavior is retained on older versions
in the interest of breaking as little as possible.  Theoretically, there
shouldn't be a problem with just always using pixel units, but I don't
want to investigate this if I don't have to.

TL;DR:  With MATLAB R2015b, a pixel is no longer represents the same
fixed fraction of an inch independent of the display size, so if we want
our plot labels to be a certain size in pixels, we should request the
font size in pixels, not points.