Set DefaultTextFontUnits and DefaultTextFontSize for examples and the Guide.
For some reason, if you specify a font size when making, say, a plot
title, it doesn't inherit the font units from the axes with which it's
associated.  Instead, it gets them from DefaultTextFontUnits.  This
behavior is a little strange because it _does_ do so when the font size
is _not_ specified.

In more detail, suppose DefaultAxesFontSize and DefaultTextFontSize are
both 10, but DefaultAxesFontUnits is 'pixels' while DefaultTextFontUnits
is 'points'.  If you call title('My title'), the title will be rendered
in a 10-pixel font, per the DefaultAxes* settings, but if you call
title('My title', 'FontSize', 16), it will be rendered in a 16-point
font, using the DefaultTextFontUnits and overriding the
DefaultTextFontSize.  Per the new "DPI-awareness" rules, the system will
try to render that in an actual 16-point font.  That requires a lot of
pixels on high-DPI displays, and that's bad for web graphics, which have
strict pixel size requirements.

Whatever.  The solution to this problem is obvious:  we need to set
DefaultTextFontSize and DefaultTextFontUnits, at least for R2015b and
later, for which all this stuff starts to matter.

Actually, the story being told in this and the previous commit is more
complicated that suggested previous because it seems that that
"DPI-awareness" in R2015b works in a different manner from what I
thought.  The resolution at which MATLAB renders figures is controlled
by the ScreenPixelsPerInch property of the graphics root.  Here is the
R2015b documentation on this property:

  Display resolution, returned as a scalar in pixels per inch. The value
  depends on the system.

      * On Windows systems, the value is 96 DPI.
      * On Macintosh systems, the value is 72 DPI.
      * On Linux system, the value is determined by your system resolution.

So, on Linux, ScreenPixelsPerInch is now supposedly the true resolution
in DPI (which it definitely is for my laptop), but on Windows and Mac,
it's 96 and 72 DPI, irrespective of the display?!  Isn't the latter the
polar _opposite_ of being "DPI-aware"?  Or perhaps MATLAB does some
internal "tricks" that make high-resolution displays on those systems
"act" from the perspective of the rest of the graphics subsystem as
though they had the prescribed resolutions?  I have no idea.

I doubt this is the last commit we'll have on this issue...