Override MATLAB built-in [xyz]label() for examples and the Guide.
The last commit fixed things as far as plot titles are concerned, but it
turns out that axis labels are yet a different story.  Setting
DefaultTextFontUnits is not enough for them because, apparently, the
built-in [xyz]label() functions manually set FontUnitsMode and
FontSizeMode to 'auto' for axis labels.  This has the effect of
overriding any manually-specified defaults for the font units.

Regrettably, I know no way around this other than to make custom
[xyz]label() functions that shadow MATLAB's built-in ones and don't have
this behavior, so that's what I've done.