Set default figure color to white in make_example.m.
Setting it to 'none' causes some very strange issues for me (running
R2015b on Linux), mostly related to plots involving chebfun2.
Specifically, it results in lines appearing on the surfaces when there
are none.  As changing it to white doesn't seem to break anything, I
think we're OK, but we should inspect plots carefully to be sure.