More fixes to example figure formats for R2015b+ "DPI-aware" behavior.
Now that I've had a chance to play with this on a new Mac running R2015b
(c/o Dr. Hashemi), I think I have finally managed to figure out what's
going on here and, more importantly, how to fix it.

R2015b brings with it a new notion of "device-independent" pixels as
opposed to "true" or "display" pixels, and all measurements in "pixels"
are now interpreted to be "device-independent" pixels instead of "true"
ones.  We need publish() to create figures at a fixed size in "true"
pixels for the website, so to fix things, we need to convert our sizes
in "true" pixels to ones in "device-independent" pixels to get things to
work.

Now, what exactly is meant by a "device-independent" pixel depends on
the platform on which MATLAB is run.  On Windows and Mac,
"device-independent" pixels are specified to be 1/96 of an inch and 1/72
of an inch, respectively, and the number of "true" pixels to which they
correspond is determined by the display resolution in "true" pixels per
inch (dpi), presumably as presented to MATLAB by the OS device driver.
On Linux, a "device-independent" pixel is the same as a "true" pixel.
Thus, we see that the conversion factor from "device-independent" to
"true" pixels depends on the screen resolution in "true" dpi, at least
on Windows and Mac.

Here's the issue:  there is no "straightforward" way to get at the
resolution in "true" dpi.  You can easily get the screen resolution in
"device-independent" dpi with get(0, 'ScreenPixelsPerInch'), but the
resolution in "true" dpi isn't directly available (unless you're on
Linux, in which case that command also gives the resolution in "true"
dpi because "true" and "device-independent" dpi are the same on that
platform).

To get around this, the getTrueDPI() function introduced in this commit
uses print() to generate RGB data for figure that's 1" x 1", sampled "at
the screen resolution".  The resolution used by print() in this case is
the resolution in "true" dpi, which makes sense, since raw image data is
always specified in "true" pixels, not "device-independent" ones, and so
in order to get a figure that has 1" dimensions, you must use a number
of pixels equal to the screen resolution in "true" dpi.  Therefore, we
can get the screen resolution in "true" dpi simply by looking at the
size of the output.

Once the resolution in "true" dpi is in hand, converting our
measurements specified in "true" pixels to ones specified in
"device-independent" pixels is a piece of cake.