Put the original example folder on the path when publishing.
This way if an example has an external file, e.g., a *.mat file with
some data in it, it will be on the path when we run it.

This is currently needed for a temporary spherefun example, but it may
not be necessary if we ultimately decide not to allow external data
files in the future.