Adjust parsing of underscores in Markdown --> HTML renderer.
The part of the renderer responsible for handling underscore emphasis
doesn't always do the right thing when we're writing equations.  For
instance, if we write

 $$ u_t = -0.5(u^2)_x - u_{xxx}. $$

the renderer will interpret the "_x - u_" bit as text that needs to be
emphasized and turn it into "<em>x - u</em>", which makes a colossal
mess.

One can, of course, fix this by backslash-escaping the underscores, but
then the document won't compile under MATLAB publish(), since LaTeX
doesn't recognize backslash escapes.

This commit uses the Python-Markdown extensions facility to adjust the
renderer to try and fix problems like this.  The fix I've implemented
just makes it pickier about what characters it will allow to precede and
follow emphasis underscores.  This is meant to be a simple and easy fix
that will make things work in 90% of cases.

A more robust fix would be to patch the escaping mechanism to make the
Markdown renderer ignore blocks enclosed in dollar signs or double
dollar signs.  While I believe it is possible to construct a
Python-Markdown extension that will do this, it's more complicated than
I have time to bother with at the moment, so I'm going to leave it for
now.  An intermediate fix might be to at least get such blocks ignored
for emphasis underscores.