More tweaks to the underscore emphasis replacement regex.
The last version of the regex was too restrictive and had some issues
with matching emphasis underscores occurring at the end of a line (among
other things).  I think part of this may have to do with the fact that
the regex gets implicitly surrounded by "^(.*?)" and "(.*?)$" by the
Python-Markdown Pattern class and that it then gets complied with
re.DOTALL set so that '.' matches newlines too.  I was not aware of this
when I wrote the expression originally.

No matter:  we'll just use a less-restrictive solution.  This new
pattern is just the default one, modified to exclude ")}" from the list
of characters allowed to precede the first underscore and "({" from the
list of characters allowed to follow the second.