Remove MATLAB errors from example output in publish() stylesheet.
This is the right way to do this.  Rather than trying to strip the error
tags out of the Markdown document by pretnding it's HTML and "parsing"
it with libxml (see commit e0ac9164), the error tags now simply don't
get emitted by publish() in the first place.

This has the added benefit of removing the dependency of the web scripts
on Perl and libxml, though this never seemed to be much of an issue for
us in the first place.

On another note, the indentation in the XSL files is _terrible_.  (More
specifically, it's nonexistent.)  I really hope the original versions of
these were auto-generated by some tool.  You'd have to be crazy to
indent them like this if you're editing them manually...