Try to fix some issues with sorting of examples.
We've been having a minor problem with the sorting routine jumbling the
order in which the examples are listed on their respective index pages
and within the inventory.txt files.  They're always sorted correctly
according to date, but after that, they appear in any order, even though
the code apparently instructs them to sort in a specific way (l. 384:
sort e['date']+e['title'] in reverse "alphanum" order).  Moreover, the
results seem to differ from machine to machine.

This isn't really that important, but it results in a lot of unnecessary
changes each time the site is built and can cause confusion as, e.g.,
the index for examples/approx will get reordered if you've just
published a new example in examples/cheb.  This makes it harder to tell
what files have _actually_ been changed, as it clutters the output of
"git status".

The problem is rooted in the fancy zip-map implementation of
sort_alphanum().  What goes wrong is best explained by example.
Consider the following two examples in the approx3 category:

  "Loosening the Chebfun3 tolerance"          (2016-06-01)
  "The theorems of Gauss, Green, and Stokes"  (2016-06-01)

According to the specified sort order, the second of these should appear
first on the approx3 index page.  The sort key that gets generated for
the first is

  >>> s1 = "2016-06-01Loosening the Chebfun3 tolerance"
  >>> k1 = zip(re.split('(\d+)', s1.lower())[0::2], map(int, re.split('(\d+)', s1.lower())[1::2]))
  >>> k1
  [('', 2016), ('-', 6), ('-', 1), ('loosening the chebfun', 3)]

Notice that we've lost the word "tolerance" in the list of tuples that
is being compared.  This happens because the two arrays passed to zip()
don't have the same size, so the longer one gets truncated before
forming the zipped list.  For the second, the problem is even worse:

  >>> s2 = '2016-06-01The theorems of Gauss, Green and Stokes'
  >>> k2 = zip(re.split('(\d+)', s2.lower())[0::2], map(int, re.split('(\d+)', s2.lower())[1::2]))
  >>> k2
  [('', 2016), ('-', 6), ('-', 1)]

Now the title doesn't figure into the sort key at all.  Only the date
matters.  If we sort the keys in reverse, we get

  >>> sorted([k1, k2], reverse=True)
  [[('', 2016), ('-', 6), ('-', 1), ('loosening the chebfun', 3)], [('', 2016), ('-', 6), ('-', 1)]]

So the first example will appear first on the page, contrary to what was
expected.

It's clear from this example that the fancy implementation of
sort_alphanum() doesn't actually work.  (It works for the simple
examples listed on the webpage from which it was taken but not more
generally.)  This commit fixes this by using the longer-but-correct (and
much easier to understand) implementation given at the top of the page.
This implementation is still not perfect (For instance, it sorts the
string "approx2/Alignment" ahead of "approx/AbsoluteValue".  I don't
think this is correct:  '/' comes before '2' in ASCII.), but it should
be enough at least to stabilize the sort order so that we don't see so
many spurious changes to the examples indices.

This commit also sorts the examplees and Guide chapters before writing
their respective inventories, so we should see spurious changes in those
go away as well.