Don't interpret exponentiated asterisks in LaTeX formulae as Markdown.
This commit modifies the regex used by the Markdown interpreter for
detecting words *emphasized* with asterisks to ignore asterisks that are
immediately preceded by a caret ('^').  This allows you to write $A^*$
for the Hermitian adjoint of A without running into trouble.

Note that the default regex actually works most of the time; it's only
when * is applied to two matrices that appear in the same equation on
the same line that we run into trouble (e.g., $C = A^* + B^*$ would be
problematic).

Like the other hack used to modify the behavior of underscore
_emphasis_, this does NOT actually check that we're in math mode; it'll
ignore "^*" in text mode too.  This is likely not much of an issue...