Replace some HTTP URLs in base template with HTTPS ones.
Without this, we run into issues with MathJaX failing to work when the
website is viewed over HTTPS due to issues with mixed content (i.e., the
simultaneous presence of secure and non-secure items on the page).

This issue was originally spotted by LNT; the fix is from Silviu Filip.